Convert project to esm
In this diff I migrated the project to esm and provided module field
which is useful for bundlers.

My usecase is rollup. It is able to handle only esm. And this package
requires me to add commonjs and json plugins.

As a side bonus this module can be used in browsers without bundlers
similar to this https://unpkg.com/jss@next?module.

For `tap` I wasn't able to setup esm so I replaced it with `tape`.